Set colour theme in VS Code settings
I have different colour themes per project so can visually pick out
which window is which.

I don't really wanna check this into the project but it seems like
there's no other easy way.

If you are editing the project in VS Code and this setting is annoying
for you, please feel free to just revert this commit.