Lock userspace allocator threads to CPUs
Reiji pointed out this is likely to cause unwanted variance.

From a quick test it seems we can get just as good performance with this
logic. To keep the maths simple the mmapSize variable is now calculated
based on the number of CPUs. That seems fine since there's no good
reason to compare runs of this benchmark between different machines.