Ensmallen kernel allocation set
I'm seeing instability in the workload, sometimes the OOM killer gets too
enthusiastic and kills the parent thread as well as the child.. Experimentally,
reducing the amount of kernel allocations seems to fix this although I don't
know why.

It no longer makes sense to shift by the allocation orders since the allocations
are small again now, and that would cause `target` to underflow when order is
large.