Support overriding makeself path
Not sure what's going on but I'm getting issues with Makself, see
https://github.com/megastep/makeself/issues/324.

Haven't really figured out the details so I'm not gonna document this in the
README for now.

My workaround is to checkout commit 5809be74 of makeself, build it (just `make`),
then run the `release/makeself-2.4.5.run` it generates, and use the `makeself.sh`
that I find in its output directory. Not sure why you can't just run the
`makeself-2.4.5.run` directly.