Use reservoir sampling for latencies
I just discovered this simple technique for maintaining a uniform
sample of a population without knowing the population's size ahead of
time.

This helps allay fears that the results are biased due to only sampling
towards the end of the kallocfree workload's run.

I believe the use of generics here is slightly more enthusiastic than
Google recommends - we are not supposed to use it unless there are
genuinely different types being used for type-parameters. However in
this case I think having references to time.Duration in sampling.go
would be a confusing misdirection. It's helpful that the type of the
data being sampled is unknown because it makes it very obvious that the
sampling algorithm is not looking at the data itself.